# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Login user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refreshToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token"
          }
        },
        "summary": "Refresh access token using refresh token",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully logged out"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Logout user (revoke current token)",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/recovery-password/{userId}": {
      "post": {
        "operationId": "AuthController_recoveryPassword",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password recovery token generated"
          },
          "403": {
            "description": "Forbidden - Admin role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Request password recovery token (Admin only)",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully"
          },
          "401": {
            "description": "Invalid or expired token"
          }
        },
        "summary": "Reset password with token",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/orders": {
      "get": {
        "operationId": "OrdersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all orders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all orders",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Order ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "404": {
            "description": "Order not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get order by ID",
        "tags": [
          "orders"
        ]
      },
      "patch": {
        "operationId": "OrdersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Order ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "404": {
            "description": "Order not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update order",
        "tags": [
          "orders"
        ]
      }
    },
    "/groups": {
      "post": {
        "operationId": "GroupsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Group created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin role required"
          },
          "409": {
            "description": "Group with this name already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new group",
        "tags": [
          "Groups"
        ]
      },
      "get": {
        "operationId": "GroupsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Groups retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Group"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all groups",
        "tags": [
          "Groups"
        ]
      }
    },
    "/comments": {
      "post": {
        "operationId": "CommentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment created successfully. User assigned to order on first comment.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "400": {
            "description": "Order not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a comment for an order",
        "tags": [
          "Comments"
        ]
      },
      "get": {
        "operationId": "CommentsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Comments retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Comment"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all comments",
        "tags": [
          "Comments"
        ]
      }
    }
  },
  "info": {
    "title": "CRM Programming School API",
    "description": "API for managing CRM Programming School, teachers, and courses",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "password123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT access token"
          },
          "refreshToken": {
            "type": "string",
            "description": "JWT refresh token"
          },
          "accessExpiresAt": {
            "format": "date-time",
            "type": "string",
            "description": "Access token expiration date"
          },
          "refreshExpiresAt": {
            "format": "date-time",
            "type": "string",
            "description": "Refresh token expiration date"
          },
          "user": {
            "type": "object",
            "description": "User information"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "accessExpiresAt",
          "refreshExpiresAt",
          "user"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Reset token"
          },
          "newPassword": {
            "type": "string",
            "description": "New password",
            "minLength": 6
          }
        },
        "required": [
          "token",
          "newPassword"
        ]
      },
      "Group": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Group ID"
          },
          "name": {
            "type": "string",
            "description": "Group name",
            "maxLength": 255
          },
          "orders": {
            "description": "Related orders",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation date"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update date"
          }
        },
        "required": [
          "id",
          "name",
          "orders",
          "createdAt",
          "updatedAt"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "User ID"
          },
          "email": {
            "type": "string",
            "description": "User email"
          },
          "firstName": {
            "type": "string",
            "description": "User first name"
          },
          "lastName": {
            "type": "string",
            "description": "User last name"
          },
          "role": {
            "type": "string",
            "description": "User role",
            "enum": [
              "admin",
              "manager"
            ]
          },
          "recoveryPasswordToken": {
            "type": "string",
            "description": "Password recovery token"
          },
          "recoveryPasswordExpires": {
            "format": "date-time",
            "type": "string",
            "description": "Password recovery token expiry"
          },
          "is_active": {
            "type": "boolean",
            "description": "Account activation status"
          },
          "tokens": {
            "description": "User tokens",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "managedOrders": {
            "description": "Managed orders",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "comments": {
            "description": "User comments",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation date"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update date"
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "role",
          "recoveryPasswordToken",
          "recoveryPasswordExpires",
          "is_active",
          "tokens",
          "managedOrders",
          "comments",
          "createdAt",
          "updatedAt"
        ]
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Order ID"
          },
          "name": {
            "type": "string",
            "description": "Student first name",
            "maxLength": 25
          },
          "surname": {
            "type": "string",
            "description": "Student surname",
            "maxLength": 25
          },
          "email": {
            "type": "string",
            "description": "Student email",
            "maxLength": 100
          },
          "phone": {
            "type": "string",
            "description": "Student phone",
            "maxLength": 12
          },
          "age": {
            "type": "number",
            "description": "Student age"
          },
          "course": {
            "type": "string",
            "description": "Course",
            "enum": [
              "FS",
              "QACX",
              "JCX",
              "JSCX",
              "FE",
              "PCX"
            ]
          },
          "course_format": {
            "type": "string",
            "description": "Course format",
            "enum": [
              "static",
              "online"
            ]
          },
          "course_type": {
            "type": "string",
            "description": "Course type",
            "enum": [
              "pro",
              "minimal",
              "premium",
              "incubator",
              "vip"
            ]
          },
          "sum": {
            "type": "number",
            "description": "Total sum"
          },
          "alreadyPaid": {
            "type": "number",
            "description": "Already paid amount"
          },
          "utm": {
            "type": "string",
            "description": "UTM parameters",
            "maxLength": 100
          },
          "msg": {
            "type": "string",
            "description": "Message",
            "maxLength": 100
          },
          "status": {
            "type": "string",
            "description": "Order status",
            "enum": [
              "In work",
              "New",
              "Aggre",
              "Disaggre",
              "Dubbing"
            ]
          },
          "group": {
            "description": "Related group",
            "allOf": [
              {
                "$ref": "#/components/schemas/Group"
              }
            ]
          },
          "groupId": {
            "type": "number",
            "description": "Group ID"
          },
          "comments": {
            "description": "Order comments",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "manager": {
            "description": "Assigned manager",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "managerId": {
            "type": "number",
            "description": "Manager ID"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "Creation date"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "description": "Last update date"
          }
        },
        "required": [
          "id",
          "name",
          "surname",
          "email",
          "phone",
          "age",
          "course",
          "course_format",
          "course_type",
          "sum",
          "alreadyPaid",
          "utm",
          "msg",
          "status",
          "group",
          "groupId",
          "comments",
          "manager",
          "managerId",
          "created_at",
          "updated_at"
        ]
      },
      "UpdateOrderDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Student first name",
            "maxLength": 25
          },
          "surname": {
            "type": "string",
            "description": "Student surname",
            "maxLength": 25
          },
          "email": {
            "type": "string",
            "description": "Student email",
            "maxLength": 100
          },
          "phone": {
            "type": "string",
            "description": "Student phone",
            "maxLength": 12
          },
          "age": {
            "type": "number",
            "description": "Student age"
          },
          "course": {
            "type": "string",
            "description": "Course",
            "enum": [
              "FS",
              "QACX",
              "JCX",
              "JSCX",
              "FE",
              "PCX"
            ]
          },
          "course_format": {
            "type": "string",
            "description": "Course format",
            "enum": [
              "static",
              "online"
            ]
          },
          "course_type": {
            "type": "string",
            "description": "Course type",
            "enum": [
              "pro",
              "minimal",
              "premium",
              "incubator",
              "vip"
            ]
          },
          "sum": {
            "type": "number",
            "description": "Total sum"
          },
          "alreadyPaid": {
            "type": "number",
            "description": "Already paid amount"
          },
          "utm": {
            "type": "string",
            "description": "UTM parameters",
            "maxLength": 100
          },
          "msg": {
            "type": "string",
            "description": "Message",
            "maxLength": 100
          },
          "status": {
            "type": "string",
            "description": "Order status",
            "enum": [
              "In work",
              "New",
              "Aggre",
              "Disaggre",
              "Dubbing"
            ]
          },
          "groupId": {
            "type": "number",
            "description": "Group ID"
          }
        }
      },
      "CreateGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Group name",
            "example": "Frontend Development",
            "maxLength": 255
          }
        },
        "required": [
          "name"
        ]
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Comment message",
            "example": "This order needs additional review"
          },
          "orderId": {
            "type": "number",
            "description": "Order ID to comment on",
            "example": 1
          }
        },
        "required": [
          "message",
          "orderId"
        ]
      },
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Comment ID"
          },
          "message": {
            "type": "string",
            "description": "Comment message"
          },
          "user": {
            "description": "User who created the comment",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "userId": {
            "type": "number",
            "description": "User ID"
          },
          "order": {
            "description": "Related order",
            "allOf": [
              {
                "$ref": "#/components/schemas/Order"
              }
            ]
          },
          "orderId": {
            "type": "number",
            "description": "Order ID"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation date"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update date"
          }
        },
        "required": [
          "id",
          "message",
          "user",
          "userId",
          "order",
          "orderId",
          "createdAt",
          "updatedAt"
        ]
      }
    }
  }
}